minor refactors, cleanup folders
