--- src/main/java/at/ac/fhcampuswien/richAF/view/Dashboard.java
@@ -1,21 +1,0 @@
-package at.ac.fhcampuswien.richAF.view;
-
-import javafx.application.Application;
-import javafx.geometry.Insets;
-import javafx.scene.Scene;
-import javafx.scene.control.Button;
-import javafx.scene.control.Label;
-import javafx.scene.layout.HBox;
-import javafx.scene.layout.VBox;
-import javafx.stage.Stage;
-
-public class Dashboard extends Application {
-
-    @Override
-    public void start(Stage stage) {
-    }
-
-    public static void main(String[] args) {
-        launch();
-    }
-}
